Document modular TUI refactor
