+++ vtcode-core/src/ui/mod.rs
@@ -5,12 +5,12 @@
 
 pub mod diff_renderer;
 pub mod markdown;
-pub mod ratatui;
 pub mod slash;
 pub mod spinner;
 pub mod styled;
 pub mod terminal;
 pub mod theme;
+pub mod tui;
 pub mod user_confirmation;
 
 pub use markdown::*;
@@ -19,6 +19,7 @@ pub use spinner::*;
 pub use styled::*;
 pub use terminal::*;
 pub use theme::*;
+pub use tui::*;
 
 #[cfg(test)]
 mod tests {
